+++ src/hri_planner/component_test.cpp
@@ -24,22 +24,22 @@
 bool test_belief_update(hri_planner::BeliefUpdate::Request& req,
                         hri_planner::BeliefUpdate::Response& res) {
     // extract the messages
-    Eigen::Map<Eigen::VectorXf> xr(req.xr.data(), req.xr.size());
-    Eigen::Map<Eigen::VectorXf> ur(req.ur.data(), req.ur.size());
-    Eigen::Map<Eigen::VectorXf> xh0(req.xh0.data(), req.xh0.size());
+    Eigen::Map<Eigen::VectorXd> xr(req.xr.data(), req.xr.size());
+    Eigen::Map<Eigen::VectorXd> ur(req.ur.data(), req.ur.size());
+    Eigen::Map<Eigen::VectorXd> xh0(req.xh0.data(), req.xh0.size());
 
     // construct the shared configuration
     auto config = std::make_shared<hri_planner::SharedConfig>();
 
     // construct the belief update object
-    float ratio;
-    float decay_rate;
-    std::vector<float> fcorrection(2, 0);
+    double ratio;
+    double decay_rate;
+    std::vector<double> fcorrection(2, 0);
 
-    ros::param::param<float>("~explicit_comm/ratio", ratio, 100.0f);
-    ros::param::param<float>("~explicit_comm/decay_rate", decay_rate, 2.5f);
-    ros::param::param<float>("~explicit_comm/fcorrection_hp", fcorrection[hri_planner::HumanPriority], 3.0);
-    ros::param::param<float>("~explicit_comm/fcorrection_rp", fcorrection[hri_planner::RobotPriority], 30.0);
+    ros::param::param<double>("~explicit_comm/ratio", ratio, 100.0);
+    ros::param::param<double>("~explicit_comm/decay_rate", decay_rate, 2.5);
+    ros::param::param<double>("~explicit_comm/fcorrection_hp", fcorrection[hri_planner::HumanPriority], 3.0);
+    ros::param::param<double>("~explicit_comm/fcorrection_rp", fcorrection[hri_planner::RobotPriority], 30.0);
 
     hri_planner::BeliefModelExponential belief_model(config, ratio, decay_rate, fcorrection);
 
@@ -47,7 +47,7 @@ bool test_belief_update(hri_planner::BeliefUpdate::Request& req,
 
     // compute all the beliefs
     for (int t = 0; t < req.t_total - config->T; ++t) {
-        Eigen::Vector2f belief;
+        Eigen::Vector2d belief;
         belief_model.belief_update(xr.segment(t*config->nXr, config->nXr*config->T),
                                    ur.segment(t*config->nUr, config->nUr*config->T),
                                    xh0.segment(t*config->nXh, config->nXh),
