+++ src/hri_planner/costs.cpp
@@ -14,9 +14,9 @@
 namespace hri_planner {
 
 //----------------------------------------------------------------------------------
-float LinearCost::compute(const Trajectory &robot_traj, const Trajectory &human_traj)
+double LinearCost::compute(const Trajectory &robot_traj, const Trajectory &human_traj)
 {
-    float cost = 0.0;
+    double cost = 0.0;
 
     for (int i = 0; i < nfeatures_; ++i) {
         cost += weights_[i] * features_[i]->compute(robot_traj, human_traj);
@@ -32,7 +32,7 @@ void LinearCost::grad_ur(const Trajectory &robot_traj, const Trajectory &human_t
 
     int len = human_traj.traj_control_size();
     for (int i = 0; i < nfeatures_; ++i) {
-        Eigen::VectorXf grad_f(len);
+        Eigen::VectorXd grad_f(len);
         features_[i]->grad_ur(robot_traj, human_traj, grad_f);
 
         grad += weights_[i] * grad_f;
@@ -46,7 +46,7 @@ void LinearCost::grad_uh(const Trajectory &robot_traj, const Trajectory &human_t
 
     int len = human_traj.traj_control_size();
     for (int i = 0; i < nfeatures_; ++i) {
-        Eigen::VectorXf grad_f(len);
+        Eigen::VectorXd grad_f(len);
         features_[i]->grad_uh(robot_traj, human_traj, grad_f);
 
         grad += weights_[i] * grad_f;
@@ -54,14 +54,14 @@ void LinearCost::grad_uh(const Trajectory &robot_traj, const Trajectory &human_t
 }
 
 //----------------------------------------------------------------------------------
-void LinearCost::add_feature(float weight, FeatureBase *feature)
+void LinearCost::add_feature(double weight, FeatureBase *feature)
 {
     weights_.push_back(weight);
     features_.push_back(std::shared_ptr<FeatureBase>(feature));
 }
 
 //----------------------------------------------------------------------------------
-void LinearCost::add_feature(float weight, const std::shared_ptr<FeatureBase> feature)
+void LinearCost::add_feature(double weight, const std::shared_ptr<FeatureBase> feature)
 {
     weights_.push_back(weight);
     features_.push_back(feature);
@@ -74,7 +74,7 @@ void HumanCost::hessian_uh(const Trajectory &robot_traj, const Trajectory &human
 
     int len = human_traj.traj_control_size();
     for (int i = 0; i < nfeatures_; ++i) {
-        Eigen::MatrixXf hess_f(len, len);
+        Eigen::MatrixXd hess_f(len, len);
         static_cast<FeatureHumanCost*>(features_[i].get())->hessian_uh(robot_traj, human_traj, hess_f);
 
         hess += weights_[i] * hess_f;
@@ -88,7 +88,7 @@ void HumanCost::hessian_uh_ur(const Trajectory &robot_traj, const Trajectory &hu
 
     int len = human_traj.traj_control_size();
     for (int i = 0; i < nfeatures_; ++i) {
-        Eigen::MatrixXf hess_f(len, len);
+        Eigen::MatrixXd hess_f(len, len);
         static_cast<FeatureHumanCost*>(features_[i].get())->hessian_uh_ur(robot_traj, human_traj, hess_f);
 
         hess += weights_[i] * hess_f;
@@ -102,7 +102,7 @@ void SingleTrajectoryCost::set_trajectory_data(const Trajectory &traj)
 }
 
 //----------------------------------------------------------------------------------
-float SingleTrajectoryCostRobot::compute(const Trajectory &traj)
+double SingleTrajectoryCostRobot::compute(const Trajectory &traj)
 {
     return compute(traj, const_traj_);
 }
@@ -114,7 +114,7 @@ void SingleTrajectoryCostRobot::grad(const Trajectory &traj, VecRef grad)
 }
 
 //----------------------------------------------------------------------------------
-float SingleTrajectoryCostHuman::compute(const Trajectory &traj)
+double SingleTrajectoryCostHuman::compute(const Trajectory &traj)
 {
     return compute(const_traj_, traj);
 }
